分割文件并优化
