+++ GoogleTranslateIpCheck/GoogleTranslateIpCheck.csproj
@@ -4,10 +4,28 @@
         <TargetFramework>net7.0</TargetFramework>
         <Nullable>enable</Nullable>
         <PublishAot>true</PublishAot>
+        <AnalysisLevel>latest-recommended</AnalysisLevel>
     </PropertyGroup>
     <ItemGroup>
         <PackageReference Include="Flurl.Http" Version="4.0.0-pre2" />
         <PackageReference Include="IPNetwork2" Version="2.6.556" />
+    </ItemGroup>
+    <ItemGroup>
+        <Compile Update="Properties\Text.Designer.cs">
+            <DesignTime>True</DesignTime>
+            <AutoGen>True</AutoGen>
+            <DependentUpon>Text.resx</DependentUpon>
+        </Compile>
+    </ItemGroup>
+    <ItemGroup>
+        <EmbeddedResource Update="Properties\Text.resx">
+            <Generator>ResXFileCodeGenerator</Generator>
+            <LastGenOutput>Text.Designer.cs</LastGenOutput>
+        </EmbeddedResource>
+        <EmbeddedResource Update="Resources\Text.resx">
+            <Generator>ResXFileCodeGenerator</Generator>
+            <LastGenOutput>Text.Designer.cs</LastGenOutput>
+        </EmbeddedResource>
     </ItemGroup>
     <ItemGroup>
         <None Update="config.json">
