+++ GoogleTranslateIpCheck/IpRegex.cs
@@ -2,7 +2,7 @@
 
 namespace GoogleTranslateIpCheck;
 
-public partial class IpRegex
+public static partial class IPRegex
 {
     [GeneratedRegex(@"^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])$")]
     public static partial Regex IPv4Regex( );
