+++ GoogleTranslateIpCheck/_JSONContext.cs
@@ -1,0 +1,6 @@
+﻿using System.Text.Json.Serialization;
+
+namespace GoogleTranslateIpCheck;
+
+[JsonSerializable(typeof(Config))]
+public partial class _JSONContext : JsonSerializerContext { }
